Rename and implement DI
